requirements.txt
Использованные библиотеки:
    PyQT
    sys
    sqlite